refactor: rename annotation/ to generate_dataset/, unify session files
ball_detection:
- Move youtube/candidate_workflow.py to generate_dataset/
- Rename annotation/youtube_session.py to generate_dataset/annotation_session.py
- Update generate_dataset/__init__.py with full re-exports
- Remove candidate_workflow re-exports from youtube/__init__.py
- Update all import paths (scripts, README)

court_detection:
- Rename annotation/session.py to generate_dataset/annotation_session.py
- Update __init__.py and all import paths
