+++ src/tasks/ball_detection/generate_dataset/annotation_session.py
