+++ src/tasks/ball_detection/generate_dataset/candidate_workflow.py
