+++ src/tasks/ball_detection/scripts/youtube/annotate_youtube_ball.py
@@ -21,7 +21,7 @@ import hydra
 from hydra.utils import to_absolute_path
 from omegaconf import DictConfig
 
-from src.tasks.ball_detection.annotation import (
+from src.tasks.ball_detection.generate_dataset import (
     BallAnnotationSessionConfig,
     FinalizeConfig,
     ZoomConfig,
