+++ src/tasks/court_detection/generate_dataset/__init__.py
@@ -1,6 +1,6 @@
 """Court keypoint annotation helpers."""
 
-from src.tasks.court_detection.annotation.session import (
+from src.tasks.court_detection.generate_dataset.annotation_session import (
     AnnotationSessionConfig,
     run_annotation_session,
 )
